Feat: Require authentication on existing API endpoints (#12)
* Refactor: remove circular imports and adjust imports

* Refactor: create 'Settings' class

* Feat: adjust 'domain' related functions

* refactor: centralização funcionalidades do user

* docs: correção type hint

* feat: adição associação lista produto

* feat: adição compartilhamento listas

---------

+++ src/lists/domain/use_cases.py
@@ -1,35 +1,65 @@
 from uuid import UUID
 
-from core.errors import NotFoundError
-from lists.schemas.list import ListCreate, ListRead
+from core.domain.errors import ForbiddenError, NotFoundError
+from shared.list_shares.domain.ports import ListShareRepositoryPort
+from shared.list_shares.domain.use_cases import validate_user_access_to_list
 
+from .filters import ListFilters
 from .ports import ListRepositoryPort
+from .schemas import ListCreate, ListRead
 
 
 async def create_list(
-    list_payload: ListCreate, repository: ListRepositoryPort
+    repository: ListRepositoryPort,
+    payload: ListCreate,
+    user_id: UUID,
 ) -> ListRead:
-    created = await repository.create(list_payload)
+    payload = payload.copy(update={"created_by_id": user_id})
+    return await repository.create(payload)
 
-    return created
 
-
-async def read_list(list_id: UUID, repository: ListRepositoryPort) -> ListRead:
-    list_obj = await repository.read(list_id)
-
-    if not list_obj:
-        raise NotFoundError("List", str(list_id))
+async def read_list(
+    repository: ListRepositoryPort,
+    share_repository: ListShareRepositoryPort,
+    list_id: UUID,
+    user_id: UUID,
+) -> ListRead:
+    list_obj = await validate_user_access_to_list(
+        share_repository,
+        repository,
+        list_id,
+        user_id,
+    )
 
     return list_obj
 
 
-async def read_lists(repository: ListRepositoryPort) -> list[ListRead]:
-    lists = await repository.read_all()
+async def read_lists(
+    repository: ListRepositoryPort,
+    filters: ListFilters,
+    user_id: UUID,
+) -> list[ListRead]:
+    lists = await repository.read_all_accessible(filters, user_id)
 
     return lists
 
 
-async def delete_list(list_id: UUID, repository: ListRepositoryPort) -> ListRead:
+async def delete_list(
+    repository: ListRepositoryPort,
+    share_repository: ListShareRepositoryPort,
+    list_id: UUID,
+    user_id: UUID,
+) -> ListRead:
+    list: ListRead = await read_list(
+        repository,
+        share_repository,
+        list_id,
+        user_id,
+    )
+
+    if user_id != list.created_by_id:
+        raise ForbiddenError()
+
     deleted = await repository.delete(list_id)
 
     if not deleted:
